requirements.txt
streamlit>=1.28.0,<2.0.0
langchain-google-genai>=2.0.5,<3.0.0
google-generativeai>=0.8.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0